# packages/common/src/openapibundle.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Lightdash API",
        "description": "API spec for Lightdash server",
        "contact": {
            "email": "support@lightdash.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/lightdash/lightdash/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
...
    "servers": [
        {
            "url": "http://localhost:8080/api/v1"
        }
    ],
...
    "paths": {
...
        "/user": {
            "get": {
                "summary": "Get user profile",
                "tags": ["user"],
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "Successfully found user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Success"
                                        },
                                        {
                                            "properties": {
                                                "results": {
                                                    "$ref": "#/components/schemas/LightdashUser"
                                                }
                                            },
                                            "required": ["results"]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/components/responses/ErrorResponse"
                    }
                }
            },
            "post": {
                "summary": "Create user",
                "tags": ["user"],
                "operationId": "createUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUser"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successfully created user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Success"
                                        },
                                        {
                                            "properties": {
                                                "results": {
                                                    "$ref": "#/components/schemas/LightdashUser"
                                                }
                                            },
                                            "required": ["results"]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/components/responses/ErrorResponse"
                    }
                }
            }
        },
...
        "/org/users": {
            "get": {
                "summary": "Get organization users",
                "tags": ["organization"],
                "operationId": "getOrganizationUsers",
                "responses": {
                    "200": {
                        "description": "Successfully get organization users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Success"
                                        },
                                        {
                                            "properties": {
                                                "results": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/OrganizationUser"
                                                    }
                                                }
                                            },
                                            "required": ["results"]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/components/responses/ErrorResponse"
                    }
                }
            }
        },
...
        "/org/users/{userUuid}": {
            "patch": {
                "summary": "Update organization member",
                "tags": ["organization"],
                "operationId": "updateOrganizationMember",
                "parameters": [
                    {
                        "in": "path",
                        "name": "userUuid",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "role": {
                                        "type": "string"
                                    }
                                },
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully updated user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Success"
                                        },
                                        {
                                            "$ref": "#/components/schemas/OrganizationUser"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/components/responses/ErrorResponse"
                    }
                }
            },
            "delete": {
                "summary": "Delete organization user",
                "tags": ["organization"],
                "operationId": "deleteOrganizationUser",
                "parameters": [
                    {
                        "in": "path",
                        "name": "userUuid",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/SuccessResponse"
                    },
                    "default": {
                        "$ref": "#/components/responses/ErrorResponse"
                    }
                }
            }
        },
...
        "/me/personal-access-tokens": {
            "get": {
                "summary": "Get all personal access tokens",
                "description": "List all personal access tokens for user",
                "tags": ["user"],
                "operationId": "getPersonalAccessTokens",
                "responses": {
                    "200": {
                        "description": "List all personal access tokens for user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Success"
                                        },
                                        {
                                            "properties": {
                                                "results": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/PersonalAccessToken"
                                                    }
                                                }
                                            },
                                            "required": ["results"]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/components/responses/ErrorResponse"
                    }
                }
            },
            "post": {
                "summary": "Create a personal access token",
                "tags": ["user"],
                "operationId": "createPersonalAccessToken",
                "requestBody": {
                    "description": "Personal access token",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePersonalAccessToken"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Details of personal access token including the token in plain text.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Success"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "results": {
                                                    "$ref": "#/components/schemas/PersonalAccessTokenWithSecret"
                                                }
                                            },
                                            "required": ["results"]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/components/responses/ErrorResponse"
                    }
                }
            }
        }
    },
    "security": [
        {
            "cookieAuth": []
        }
    ],
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "connect.sid"
            }
        },
        "schemas": {
...
            "Error": {
                "type": "object",
                "required": ["status", "error"],
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": ["error"]
                    },
                    "error": {
                        "type": "object",
                        "required": ["name", "statusCode", "message"],
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "statusCode": {
                                "type": "number"
                            },
                            "message": {
                                "type": "string"
                            },
                            "data": {
                                "type": "object"
                            }
                        }
                    }
                }
            },
            "Success": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": ["ok"]
                    }
                }
            },
            "LightdashUser": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "userUuid": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "organizationUuid": {
                        "type": "string"
                    },
                    "organizationName": {
                        "type": "string"
                    },
                    "isTrackingAnonymized": {
                        "type": "boolean"
                    },
                    "isMarketingOptedIn": {
                        "type": "boolean"
                    },
                    "isSetupComplete": {
                        "type": "boolean"
                    },
                    "role": {
                        "type": "string"
                    }
                },
                "required": [
                    "userUuid",
                    "firstName",
                    "lastName",
                    "organizationUuid",
                    "organizationName",
                    "isTrackingAnonymized",
                    "isMarketingOptedIn",
                    "isSetupComplete",
                    "role"
                ]
            },
            "CreateUser": {
                "type": "object",
                "properties": {
                    "inviteCode": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "inviteCode",
                    "firstName",
                    "lastName",
                    "email",
                    "password"
                ]
            },
            "OrganizationUser": {
                "type": "object",
                "properties": {
                    "userUuid": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "required": ["userUuid", "firstName", "lastName"]
            },
...
            "PersonalAccessToken": {
                "type": "object",
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "expiresAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "required": ["uuid", "createdAt", "description"]
            },
            "CreatePersonalAccessToken": {
                "type": "object",
                "properties": {
                    "expiresAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "required": ["description"]
            },
            "PersonalAccessTokenWithSecret": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/PersonalAccessToken"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "token": {
                                "type": "string"
                            }
                        },
                        "required": ["token"]
                    }
                ]
            }
        },
        "responses": {
            "ErrorResponse": {
                "description": "Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "SuccessResponse": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Success"
                        }
                    }
                }
            }
        }
    }
}